refactor: encapsulated gathering and creation of actors
