+++ src/scripts/import-actor/import-actor.js
@@ -1,0 +1,9 @@
+import { logConsole } from "../log.js";
+import { gatherActorData } from "./gather-actor-data.js";
+import { createActor } from "./create-actor.js";
+
+export async function importActor(importedActorData) {
+  const actorData = gatherActorData(importedActorData);
+  logConsole("actorData", actorData);
+  await createActor(actorData);
+}
